stuff for datastore

+++ src/request/TakeTurn.java
@@ -2,23 +2,23 @@ package request;
 
 public class TakeTurn {
 
-	private int playerID;
+	private long playerID;
 	private int currentScore;
 	
 	public TakeTurn(){
 		
 	}
 	
-	public TakeTurn(int playerID, int currentScore){
+	public TakeTurn(long playerID, int currentScore){
 		this.playerID = playerID;
 		this.currentScore = currentScore;
 	}
 
-	public int getPlayerID() {
+	public long getPlayerID() {
 		return playerID;
 	}
 
-	public void setPlayerID(int playerID) {
+	public void setPlayerID(long playerID) {
 		this.playerID = playerID;
 	}
